终身学习系统/demo/backend/requirements.txt
# FastAPI 核心
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# 数据库
sqlalchemy==2.0.25
aiosqlite==0.19.0

# 认证
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2

# 文档处理
PyPDF2==3.0.1
python-docx==1.1.0
chardet==5.2.0

# AI 和向量
openai==1.12.0
tiktoken==0.5.2
numpy==1.26.3

# 工具
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0
aiofiles==23.2.1
httpx==0.26.0
